Fix properly issue#36
The set destination method of Car class was poorly designed and is
performing two tasks in one method. Now, the check for exception raised
during illegal change in direction is handled by Car.setDestination
method and simply ignore it. It is ignored because if the destination
height couldn't be set at that instant, then the elevator keeps
monitoring its movement.